crypo 1.0 changes

+++ api-tests/ff/ipc/test_i002/test_i002.c
@@ -184,7 +184,7 @@ int32_t client_test_psa_call_with_allowed_type_values(caller_security_t caller _
 {
    int32_t            status = VAL_STATUS_SUCCESS;
    psa_handle_t       handle = 0;
-   int32_t            type[] = {PSA_IPC_CALL, 1, 2, INT32_MAX};
+   int16_t            type[] = {PSA_IPC_CALL, 1, 2, INT16_MAX};
    uint32_t           i = 0;
 
    val->print(PRINT_TEST, "[Check 5] Test psa_call with different type values\n", 0);
